\name{archetypes-generics}
\title{Generic functions in archetypes package}
\description{These generic functions are defined in the package
  archetypes.}
\alias{rss}
\usage{rss(object, ...)
nparameters(object, ...)
panorama(object, ...)
bestModel(object, ...)
xyplot(x, ...)
pcplot(x, ...)
}
\value{\code{rss}: Residual sum of squares.

\code{nparameters}: Number of parameters.


\code{panorama}: Undefined.


\code{bestModel}: The best models.


\code{xyplot}: Undefined.


}
\arguments{\item{object}{An object.}
\item{...}{Ignored.}
\item{x}{An object.}
}
\alias{nparameters}
\alias{panorama}
\alias{bestModel}
\alias{xyplot}
\alias{pcplot}

