% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-package.R
\docType{package}
\name{axe-package}
\alias{axe-package}
\alias{_PACKAGE}
\title{axe package}
\description{
This package provides a simple interface for expanding upon the `formula` class by introducing additional attributes, such as specific roles, labels, and groups to individual terms. This allows for formulas to be expanded and to create families of formulas that are related, using formula lists. This is driven by the need for more clear and intentional relationship descriptions, particularly in the causal modeling space.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/axe}
  \item Report bugs at \url{https://github.com/asshah4/axe/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
