\name{barplot.archetypes}
\alias{barplot.archetypes}
\title{Barplot of archetypes.}
\usage{\method{barplot}{archetypes} (height, data, beside=TRUE,
    percentage=FALSE, ...)}
\description{Barplot of archetypes.}
\value{Undefined.}
\arguments{\item{height}{An \code{\link{archetypes}} object.}
\item{data}{The original data matrix.}
\item{beside}{Portray the archetypes as stacked bars.}
\item{percentage}{Show real values or percentages according to the
original data.}
\item{...}{Passed to the underlying \code{\link{barplot}} call.}}
