% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-step.R
\name{bestModel.stepArchetypes}
\alias{bestModel.stepArchetypes}
\alias{bestModel.repArchetypes}
\title{Return best model}
\usage{
\method{bestModel}{stepArchetypes}(object, ...)

\method{bestModel}{repArchetypes}(object, ...)
}
\arguments{
\item{object}{An \code{archetypes} object.}

\item{...}{Ignored}
}
\description{
Return best model
}
