\name{body}
\docType{data}
\alias{body}
\title{Exploring relationships in body dimensions}
\description{
Body girth measurements and skeletal diameter measurements, as well as
age, weight, height and gender, are given for 507 physically active
individuals - 247 men and 260 women.
}
\usage{body}
\format{A data.frame containing 507 observations of 25 variables.}
\seealso{skel}
\references{
Heinz, Peterson, Johnson and Kerk. "Exploring relationships in body 
dimensions". Journal of Statistics Education, 11(2). 
http://www.amstat.org/publications/jse/v11n2/datasets.heinz.html
}
\keyword{datasets}
