% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cast}
\alias{cast}
\alias{cast.list_of_formulas}
\title{Cast a formula-based object as a data frame}
\usage{
cast(x, ...)

\method{cast}{list_of_formulas}(x, ...)
}
\value{
A \code{data.frame} object that is an explosion of the underlying object,
giving out the underlying terms, patterns, etc, corresponding to the
initial properties of the object formula
}
\description{
Cast a formula-based object as a data frame
}
