% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{check}
\alias{check}
\alias{is_term}
\alias{is_script}
\alias{is_formula}
\title{Identification of formula and formula-adjacent objects}
\usage{
is_term(x)

is_script(x)

is_formula(x)
}
\arguments{
\item{x}{Confirmation of an object being of the following classes:
\itemize{
\item \code{term}
\item \code{rx}
}}
}
\description{
Identification of formula and formula-adjacent objects
}
