% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{construct_script}
\alias{construct_script}
\title{Carrying out the instructions of a \code{script}}
\usage{
construct_script(rx)
}
\arguments{
\item{rx}{A \code{script} object from the \code{prescribe()} or \code{rx()} functions}
}
\description{
This function uses the components of a \code{script} object to expand a list of
formulas that can subsequently be used.
}
