% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{deparse_formula}
\alias{deparse_formula}
\title{Decompose a \code{rx} formula}
\usage{
deparse_formula(f, ...)
}
\value{
A named list of variables under their specific roles
}
\description{
This function takes a formula that has been labeled roles and parses it through to
identify the individual roles of the terms. The roles are set via the
\code{\link[rx:options]{rx::set_rx_roles()}} function, which modifies global options. The default
presumes the that left-hand side of the equation will be roleed \strong{lhs} and
the right-hand side will be roleed \strong{rhs}.
}
\keyword{internal}
