% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{divide}
\alias{divide}
\title{Divide a Formula into Parts}
\usage{
divide(x, side, ...)
}
\arguments{
\item{x}{An object of \code{rx} or \code{formula} class (or can be coerced into one)}

\item{side}{Which side of the formula should be divided into parts. This
parameter can is a character vector from \code{c("lhs", "rhs")}, and defines how
this formula division should occur.}
}
\value{
A list of \code{rx} objects
}
\description{
An \code{rx} formula can be split into simpler parts as needed. If there are
multiple dependent variables, then they can be split into individual
formulas. If there are multiple independent variables, they can also be split
into individual formulas.
}
