% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{explode}
\alias{explode}
\alias{explode.list_of_formulas}
\title{Explode out a formula-based object}
\usage{
explode(x, ...)

\method{explode}{list_of_formulas}(x, ...)
}
\value{
A \code{data.frame} object that is an explosion of the underlying
object, giving out the underlying terms, patterns, etc, corresponding to the initial properties of the object
formula
}
\description{
Explode out a formula-based object
}
