\name{[.stepArchetypes}
\alias{[.stepArchetypes}
\title{Extract method}
\usage{
  \method{[}{stepArchetypes} (x, i)
}
\arguments{
  \item{x}{A \code{stepArchetypes} object.}

  \item{i}{The indizes to extract.}
}
\value{
  A \code{stepArchetypes} object containing only the parts
  defined in \code{i}.
}
\description{
  An extraction on a \code{stepArchetypes} object returns
  again a \code{stepArchetypes} object.
}

