% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-class.R
\name{fitted.archetypes}
\alias{fitted.archetypes}
\title{Return fitted data}
\usage{
\method{fitted}{archetypes}(object, ...)
}
\arguments{
\item{object}{An \code{archetypes} object.}

\item{...}{Ignored.}
}
\value{
Matrix with approximated data.
}
\description{
Returns the approximated data.
}
