% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{formula_script}
\alias{formula_script}
\alias{prescribe}
\alias{prescribe.formula}
\alias{prescribe.term_archetype}
\alias{prescribe.default}
\alias{rx}
\title{Prescriptions}
\usage{
prescribe(x = unspecified(), ...)

\method{prescribe}{formula}(x, role = list(), group = list(), label = list(), ...)

\method{prescribe}{term_archetype}(x, role = list(), group = list(), label = list(), ...)

\method{prescribe}{default}(x = unspecified(), ...)

rx(x = unspecified(), ...)
}
\arguments{
\item{x}{Objects of the following types can be used as inputs
\itemize{
\item \code{term_archetype}
}}

\item{...}{Arguments to be passed to or from other methods}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure}: predictor variable that serves as a primary or key
variable in the \eqn{exposure -> outcome} relationship
\item \strong{outcome}: outcome/dependent variable that serves as an individual
variable in the \eqn{exposure -> outcome} relationship
\item \strong{covariate}: predictor variable that is used to adjust/control for an
additional primary variable, such as \eqn{outcome <- exposure + covariate}
\item \strong{predictor}: predictor variable that is non-specific but decidedly on
the right-hand side of an equation
\item \strong{mediator}: predictor variable that is thought to be a causal
interm_archetypeediary in the \eqn{exposure -> mediator -> outcome} pathway
\item \strong{confounder}: predictor variable that is thought to be a confounder of
the causal relationship in the \eqn{exposure <- confounder -> outcome}
pathway
\item \strong{unknown}: default role of a variable that has not yet been assigned a
place, such as a potential interm_archetypeediary object
}}

\item{group}{Grouping variable name for independent variables for modeling
terms together}

\item{label}{Display-quality label describing the variable}
}
\value{
An object of class \code{formula_script}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function defines a new modified \code{formula} class that has been
vectorized. It expands upon the functionality of formulas.
}
\section{Roles}{


Specific roles the variable plays within the formula. These are of particular
importance, as they serve as special terms that can effect how a formula is
interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure} or \code{X(...)}: a predictor variable that serves as a primary or
key variable in the \eqn{Exposure ~ Outcome} relationship
}

Formulas can be condensed by applying their specific role to individual terms
as a function/wrapper. For example, \code{y ~ X(x1) + x2 + x3}. This would signify
that \code{x1} has the specific role of an exposure.
}

\section{Pluralized Arguments}{


For the arguments that would be dispatched for objects that are plural, e.g.
containing multiple terms such as a \code{formula} object, the input should be
wrapped within a \code{list()}.

For example, for the \strong{role} argument, it would be written:

\code{role = list(X ~ "exposure", Y ~ "outcome", M ~ "mediator", C ~ "covariate")}

This applies for all others plural objects and arguments.
}

