% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rhs}
\alias{get_rhs}
\title{Get right hand side terms. If tidy = TRUE, then return without wrappers.}
\usage{
get_rhs(f, tidy = TRUE)
}
\description{
Get right hand side terms. If tidy = TRUE, then return without wrappers.
}
\keyword{internal}
