% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.named}
\alias{is.named}
\title{Identify if an Object has been Named}
\usage{
is.named(x, ...)
}
\arguments{
\item{x}{An object to be tested}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Returns \code{TRUE} or \code{FALSE} depending on if the object has been named
}
\description{
A simple function to help identify if an object has names. If it is a \code{list}
or \code{vector}, then each element should be named.
}
