% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-class.R
\name{kappa.archetypes}
\alias{kappa.archetypes}
\title{Return kappa}
\usage{
\method{kappa}{archetypes}(z, ...)
}
\arguments{
\item{z}{An \code{archetypes} object.}

\item{...}{Ignored.}
}
\value{
A vector of kappas.
}
\description{
Return kappa
}
