% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels}
\alias{labels}
\alias{labels.rx}
\alias{set_labels}
\title{Labels for an \code{rx} Object}
\usage{
\method{labels}{rx}(object)

set_labels(object, ...)
}
\arguments{
\item{object}{An object of \code{rx} class}

\item{...}{Name-value pairs of term labels}
}
\value{
A character vector or list of such vectors from an object of the \code{rx}
class.
}
\description{
These functions serve to help manipulate labels for an \code{rx} formula. They
retrieve and manipulate term labels from the \code{rx} object if available. At
minimum, will have the names of the variables themselves.
}
