% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{labels.rx}
\alias{labels.rx}
\title{Find Labels from Object}
\usage{
\method{labels}{rx}(object, ...)
}
\arguments{
\item{object}{An object of \code{rx} class}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector or list of such vectors from an object of the \code{rx} class.
}
\description{
Retrieve the term labels from the \code{rx} object if available. At minimum, will have an \emph{outcomes} and \emph{predictors}.
}
