% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list-helpers}
\alias{list-helpers}
\alias{list_to_table}
\alias{table_to_list}
\alias{formula_to_named_list}
\alias{named_list_to_formula}
\title{Convert between lists, list-formulas, and tables}
\usage{
list_to_table(x, id = "terms", val = "ops", ...)

table_to_list(x, id = "terms", ...)

formula_to_named_list(x, ...)

named_list_to_formula(x, ...)
}
\arguments{
\item{x}{\itemize{
\item For \code{list_to_table()}: A named \code{list} object
\item For \code{table_to_list()}: A \code{data.frame} object
}}

\item{id}{Name of column that contains term_archetype}

\item{val}{Name of column that contains specific values}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Convert between lists, list-formulas, and tables
}
\details{
\code{table_to_list()}:

Takes a \code{data.frame} and uses the columns to generate a named list. This
removes the original column names, as it assumes that the data is contained
within the frame itself. It defaults to using the first column as the names
of the list.

\code{formula_to_named_list()}:

Handling of list-formula arguments. Stylistic choice to make arguments
entered in the form of a list, with each entry being a formula. The LHS will
always be the terms, and the RHS will always be the non-terms item (e.g.
group, label, role, etc).

\code{named_list_to_formula()}:

Converts a named list set to a formula pattern. The LHS is the term, and the
RHS is the value, such as a label, role, or tier.
}
