% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_lists.R
\name{list_of_formulas}
\alias{list_of_formulas}
\alias{list_of_formulas.formula_rx}
\alias{list_of_formulas.default}
\alias{fmls}
\title{Prescribed Formula Lists}
\usage{
list_of_formulas(x = list(), ...)

\method{list_of_formulas}{formula_rx}(
  x,
  name = deparse1(substitute(x)),
  pattern = character(),
  data = data.frame(),
  strata = character(),
  ...
)

\method{list_of_formulas}{default}(x = list(), ...)

fmls(x = list(), ...)
}
\arguments{
\item{pattern}{This is the expansion pattern used to decide how the
covariates will incorporated into the formulas. The options are
\code{c("direct", "sequential", "parallel")}. See the details for further
explanation.
\itemize{
\item \strong{direct}: the covariates will all be included in each formula
\item \strong{sequential}: the covariates will be added sequentially, one by one, or
by groups, as indicated
\item \strong{parallel}: the covariates or groups of covariates will be placed in
parallel
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super  that combines both the \code{list} class (and
its derivative \code{list_of}) and the \code{formula} class.
}
