% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{list_of_models}
\alias{list_of_models}
\alias{list_of_models.list}
\alias{list_of_models.list_of_formulas}
\alias{list_of_models.default}
\alias{mdls}
\title{Model Lists}
\usage{
list_of_models(x, ...)

\method{list_of_models}{list}(
  x,
  labels = list(),
  roles = list(),
  name = deparse1(substitute(x)),
  ...
)

\method{list_of_models}{list_of_formulas}(x, fn, ..., data)

\method{list_of_models}{default}(x, ...)

mdls(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super  that combines both the \code{list} class (and
its derivative \code{list_of}) and regression models
}
