% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_table}
\alias{list_to_table}
\title{Convert a List into a Table}
\usage{
list_to_table(named_list, id = "terms", ...)
}
\arguments{
\item{named_list}{A \code{list} object that is named}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{data.frame} object
}
\description{
Expands a list of character vectors into a tidy, logic table.
}
