% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lists_tbls}
\alias{lists_tbls}
\alias{list_to_table}
\alias{table_to_list}
\title{Convert between lists and tables}
\usage{
list_to_table(x, id = "terms", val = "ops", ...)

table_to_list(x, id = "terms", ...)
}
\arguments{
\item{x}{\itemize{
\item For \code{list_to_table()}: A named \code{list} object
\item For \code{table_to_list()}: A \code{data.frame} object
}}

\item{id}{Name of column that contains terms}

\item{val}{Name of column that contains specific values}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{data.frame} or \code{list} object
}
\description{
Convert between lists and tables
}
\details{
For \code{table_to_list()}:

Takes a \code{data.frame} and uses the columns to generate a named list. This
removes the original column names, as it assumes that the data is contained
within the frame itself. It defaults to using the first column as the names
of the list.
}
