% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{math}
\alias{math}
\alias{Ops.rx}
\alias{add}
\alias{subtract}
\title{Formula Operations}
\usage{
\method{Ops}{rx}(e1, e2)

add(x, y, ...)

subtract(x, y, ...)
}
\arguments{
\item{e1}{objects.}

\item{e2}{objects.}

\item{x, y}{Objects of class \code{rx}, or can be coerced to \code{rx} class}

\item{...}{Further arguments passed to or from other methods}
}
\value{
An object or list of objects of \code{rx} class

An object of \code{rx} class

An \code{rx} object
}
\description{
This is a method that extends the \code{\link[methods:S4groupGeneric]{methods::Ops()}} group generic function to
include generic operators for \code{rx} formulas. This is limited to \code{rx} formulas
as to not impact the functionality of generic \code{\link[stats:formula]{stats::formula()}} objects.

Merge two formulas together based on common terms.

Subtract the elements of one formula from that of another. This only applies
to formulas where the primary object is a superset of the secondary object.
This is limited to formulas that are \code{rx} formula objects as to inot impact
the functionality of generic \code{\link[stats:formula]{stats::formula()}} objects.
}
\concept{tools}
