% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{md}
\alias{md}
\alias{model_archetype}
\alias{model_archetype.lm}
\alias{model_archetype.glm}
\alias{model_archetype.model_fit}
\alias{model_archetype.list}
\alias{model_archetype.default}
\title{Model archetype}
\usage{
model_archetype(x = unspecified(), ...)

\method{model_archetype}{lm}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_archetype}{glm}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_archetype}{model_fit}(
  x,
  model_name = deparse(substitute(x)),
  term_labels = list(),
  term_roles = list(),
  model_label = character(),
  model_description = character(),
  ...
)

\method{model_archetype}{list}(x, term_labels = list(), term_roles = list(), ...)

\method{model_archetype}{default}(x = unspecified(), ...)

md(x = unspecified(), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
