% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{merge.rx}
\alias{merge.rx}
\title{Merge Two Formulas}
\usage{
\method{merge}{rx}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{rx}, or can be coerced to \code{rx} class}
}
\value{
An object of \code{rx} class
}
\description{
Merge two formulas together based on common terms.
}
\seealso{
Other tools: 
\code{\link{Ops.rx}()}
}
\concept{tools}
