\name{moviepcplot}
\alias{moviepcplot}
\title{Archetypes parallel coordinates plot movie.}
\usage{moviepcplot(zs, data, show=c("atypes", "adata"), ssleep=0, bsleep=0,
    ...)}
\description{Archetypes parallel coordinates plot movie.}
\value{Undefined.}
\arguments{\item{zs}{An \code{\link{archetypes}} object.}
\item{data}{The data matrix.}
\item{show}{Show archetypes or approximated data.}
\item{ssleep}{Seconds to sleep before start.}
\item{bsleep}{Seconds to sleep between each plot.}
\item{...}{Passed to underlying pcplot functions.}}
