\name{movieplot2}
\alias{movieplot2}
\title{Archetypes plot movie 2.}
\usage{movieplot2(zs, data, show=atypes, ssleep=0, bsleep=0, zas.col=2,
    zas.pch=13, old.col=rgb(1, 0.5, 0.5), ...)}
\description{Archetypes plot movie 2.}
\details{Shows the intermediate steps of the algorithm;}
\value{Undefined.}
\arguments{\item{zs}{An \code{\link{archetypes}} object.}
\item{data}{The data matrix.}
\item{show}{Shows only archetypes currently.}
\item{ssleep}{Seconds to sleep before start.}
\item{bsleep}{Seconds to sleep between each plot.}
\item{zas.col}{Color of the intermediate archetypes.}
\item{zas.pch}{Type of the intermediate archetypes points.}
\item{old.col}{Color of the archetypes on step further.}
\item{...}{Passed to underlying plot functions.}}
