% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-class.R, R/archetypes-step.R
\name{nparameters.archetypes}
\alias{nparameters.archetypes}
\alias{nparameters.stepArchetypes}
\alias{nparameters.repArchetypes}
\title{Return number of archetypes}
\usage{
\method{nparameters}{archetypes}(object, ...)

\method{nparameters}{stepArchetypes}(object, ...)

\method{nparameters}{repArchetypes}(object, ...)
}
\arguments{
\item{object}{An \code{archetypes} object.}

\item{...}{Ignored.}
}
\value{
Number of archetypes.
}
\description{
Return number of archetypes
}
