% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{Ops.rx}
\alias{Ops.rx}
\title{Formula Operations}
\usage{
\method{Ops}{rx}(e1, e2)
}
\arguments{
\item{e1}{objects.}

\item{e2}{objects.}
}
\value{
An object of \code{rx} class
}
\description{
This is a method that extends the \code{\link[methods:S4groupGeneric]{methods::Ops()}} group generic function to
include generic operators for \code{rx} formulas. This is limited to \code{rx} formulas
as to not impact the functionality of generic \code{\link[stats:formula]{stats::formula()}} objects.
}
\seealso{
Other tools: 
\code{\link{merge.rx}()},
\code{\link{separate}()}
}
\concept{tools}
