% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-class.R, R/archetypes-step.R
\name{parameters,archetypes-method}
\alias{parameters,archetypes-method}
\alias{parameters-methods}
\alias{parameters,stepArchetypes-method}
\alias{parameters,repArchetypes-method}
\title{Return fitted archetypes}
\usage{
\S4method{parameters}{archetypes}(object, ...)

\S4method{parameters}{stepArchetypes}(object, ...)

\S4method{parameters}{repArchetypes}(object, ...)
}
\arguments{
\item{object}{An \code{archetypes} object.}

\item{...}{Ignored.}
}
\value{
Matrix with \eqn{k} archetypes.
}
\description{
Return fitted archetypes
}
