\name{predict.archetypes}
\alias{predict.archetypes}
\title{Predict method for archetypal analysis}
\usage{
  \method{predict}{archetypes} (object, newdata, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{newdata}{A data frame with data for which to
  predict the alpha coefficients.}

  \item{...}{Ignored.}
}
\value{
  The predict alpha coefficients.
}
\description{
  Predict method for archetypal analysis
}

