% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{prescriptions}
\alias{prescriptions}
\alias{formula_rx}
\alias{formula_rx.term_rx}
\alias{formula_rx.default}
\alias{frx}
\title{Formula vector}
\usage{
formula_rx(x = term_rx(), ...)

\method{formula_rx}{term_rx}(
  x = term_rx(),
  roles = list(),
  groups = list(),
  pattern = character(),
  ...
)

\method{formula_rx}{default}(x, ...)

frx(x = term_rx(), ...)
}
\arguments{
\item{x}{Objects of the following types can be used as inputs
\itemize{
\item \code{term_rx}
}}

\item{...}{Arguments to be passed to or from other methods}

\item{roles}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure} or \code{X(...)}: a predictor variable that serves as a primary
or key variable in the \eqn{Exposure ~ Outcome} relationship
}

Formulas can be condensed by applying their specific role to individual
terms as a function/wrapper. For example, \code{y ~ X(x1) + x2 + x3}. This would
signify that \code{x1} has the specific role of an exposure.}

\item{groups}{List of formulas that have the term (or terms) on the LHS and
the group name on the RHS (quotations to indicate character value not
necessary). E.g. \code{list(c(x1, x2) ~ grp)}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function defines a new modified \code{formula} class that has been
vectorized. It expands upon the functionality of formulas.
}
