% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-class.R
\name{residuals.archetypes}
\alias{residuals.archetypes}
\title{Return residuals}
\usage{
\method{residuals}{archetypes}(object, ...)
}
\arguments{
\item{object}{An \code{archetypes} object.}

\item{...}{Ignored.}
}
\value{
Matrix with residuals.
}
\description{
Return residuals
}
