% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{roles}
\alias{roles}
\alias{roles.rx}
\title{Find Roles from Formula}
\usage{
roles(object)

\method{roles}{rx}(object, ...)
}
\arguments{
\item{object}{A generic object that contains a formula that has been given
roles}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector or list of such vectors from an object
}
\description{
Retrieve the term roles from a statistical object if available.
}
