% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-package.R
\docType{package}
\name{rx-package}
\alias{rx}
\alias{rx-package}
\title{rx: Prescribing New Formulas}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anish Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
