% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{rx}
\alias{rx}
\alias{rx.formula}
\alias{rx.default}
\title{Prescribing a Formula}
\usage{
rx(f, ...)

\method{rx}{formula}(f, ...)

\method{rx}{default}(f, ...)
}
\arguments{
\item{f}{An object of class \code{formula}}

\item{...}{Further arguments passed to or from other methods}
}
\value{
An \code{rx} object, which has the classes of \code{c("rx", "formula")}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function defines a new modified \code{formula} class of type \code{rx}. It expands
upon the functionality of formulas.
}
\details{
When supplying a \code{formula} object, each RHS term of the formula is
considered an outcome variable, and is analyzed as a single outcome. Each
LHS term is consider a predictor, and can be modified as below:

TODO an explanation below
}
