\name{screeplot.stepArchetypes}
\alias{screeplot.stepArchetypes}
\title{Screeplot of stepArchetypes.}
\usage{\method{screeplot}{stepArchetypes}(x, type=c("lines", "barplot"), ...)
}
\description{Screeplot of stepArchetypes.}
\details{Screeplot draws the residual sum of square curve based on the best
model of each step.}
\value{Undefined.}
\arguments{\item{x}{A \code{\link{stepArchetypes}} object.}
\item{type}{Draw lines or a barplot.}
\item{...}{Passed to underlying plot functions.}
}

