% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{script}
\alias{script}
\alias{prescribe}
\alias{rx}
\title{Prescriptions}
\usage{
prescribe(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  pattern = character(),
  ...
)

rx(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  pattern = character(),
  ...
)
}
\arguments{
\item{x}{Objects of the following types can be used as inputs
\itemize{
\item \code{term_archetype}
}}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. Please see the \emph{Roles} section below for
further details. The options for roles are as below:
\itemize{
\item \strong{outcome}: outcome/dependent variable that serves as an individual
variable in the \eqn{exposure -> outcome} relationship (DEFAULT for LHS variables)
\item \strong{predictor}: predictors of the outcomes (DEFAULT for RHS variables)
\item \strong{exposure}: predictor variable that serves as a primary or key
variable in the \eqn{exposure -> outcome} relationship
\item \strong{confounder}: predictor variable that is thought to be a confounder of
the causal relationship in the \eqn{exposure <- confounder -> outcome}
pathway, normally thought of as an adjustment or controlling variable
\item \strong{mediator}: predictor variable that is thought to be a causal
interm_archetypeediary in the \eqn{exposure -> mediator -> outcome} pathway
\item \strong{unknown}: default role of a variable that has not yet been assigned a
place, such as a potential intermediary object
}}

\item{tier}{Grouping variable names for covariates or \strong{confounders} for
modeling terms together}

\item{label}{Display-quality label describing the variable}

\item{pattern}{This is the expansion pattern used to decide how the
covariates will incorporated into the formulas. The options are
\code{c("direct", "sequential", "parallel")}. See the details for further
explanation.
\itemize{
\item \strong{direct}: the covariates will all be included in each formula
\item \strong{sequential}: the covariates will be added sequentially, one by one, or
by tiers, as indicated
\item \strong{parallel}: the covariates or tiers of covariates will be placed in
parallel
}}

\item{...}{Arguments to be passed to or from other methods}
}
\value{
An object of class \code{script}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function defines a new modified \code{formula} class that has been
vectorized. It expands upon the functionality of formulas.
}
\section{Roles}{


Specific roles the variable plays within the formula. These are of particular
importance, as they serve as special terms that can effect how a formula is
interpreted. The specialized options for roles are as below:
\itemize{
\item \strong{exposure} or \code{X(...)}
\item \strong{outcome} or \code{O(...)} or placement of variable on LHS of formula
\item \strong{confounder} or \code{C(...)}
\item \strong{mediator} or \code{M(...)}
\item \strong{strata} or \code{S(...)}
}

Formulas can be condensed by applying their specific role to individual terms
as a function/wrapper. For example, \code{y ~ X(x1) + x2 + x3}. This would signify
that \code{x1} has the specific role of an exposure.
}

\section{Patterns}{


The expansion pattern allows for instructions on how the covariates should be
included in different formulas. Below, assuming that \emph{x1}, \emph{x2}, and \emph{x3} are
covariates...

\deqn{y ~ x1 + x2 + x3}

\strong{Direct}:

\deqn{y ~ x1 + x2 + x3}

\strong{Seqential}:

\deqn{y ~ x1}
\deqn{y ~ x1 + x2}
\deqn{y ~ x1 + x2 + x3}

\strong{Parallel}:

\deqn{y ~ x1}
\deqn{y ~ x2}
\deqn{y ~ x3}
}

\section{Pluralized Arguments}{


For the arguments that would be dispatched for objects that are plural,
e.g. containing multiple terms such as a \code{formula} object, the input should
be wrapped within a \code{list()}.

For example, for the \strong{role} argument, it would be written:

\code{role = list(X ~ "exposure", M ~ "mediator", C ~ "confounder")}

This applies for all others plural objects and arguments.
}

