% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx-tools.R
\name{separate}
\alias{separate}
\title{Split Formula by Another Formula}
\usage{
separate(x, y, ...)
}
\arguments{
\item{x}{Objects of class \code{rx}, or can be coerced to \code{rx} class}

\item{y}{Objects of class \code{rx}, or can be coerced to \code{rx} class}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Subtract the elements of one formula from that of another. This only applies
to formulas where the primary object is a superset of the secondary object.
This is limited to formulas that are \code{rx} formula objects as to inot impact
the functionality of generic \code{\link[stats:formula]{stats::formula()}} objects.
}
\seealso{
Other tools: 
\code{\link{Ops.rx}()},
\code{\link{merge.rx}()}
}
\concept{tools}
