% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setters}
\alias{setters}
\alias{set_roles}
\alias{set_tiers}
\alias{set_labels}
\title{Set components of term_archetype}
\usage{
set_roles(x, roles, ...)

set_tiers(x, tiers, ...)

set_labels(x, labels, ...)
}
\value{
A modified
}
\description{
Set components of term_archetype
}
