% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sides}
\alias{sides}
\alias{lhs}
\alias{rhs}
\alias{rhs.term_archetype}
\alias{lhs.term_archetype}
\alias{rhs.formula}
\alias{lhs.formula}
\alias{rhs.formula_script}
\alias{lhs.formula_script}
\title{Tools for working with formula-like objects}
\usage{
lhs(x, ...)

rhs(x, ...)

\method{rhs}{term_archetype}(x, ...)

\method{lhs}{term_archetype}(x, ...)

\method{rhs}{formula}(x, tidy = FALSE, ...)

\method{lhs}{formula}(x, tidy = FALSE, ...)

\method{rhs}{formula_script}(x, ...)

\method{lhs}{formula_script}(x, ...)
}
\arguments{
\item{tidy}{Logical value to decide if operations should be removed from the
terms. If \code{FALSE}, then the operations will remain included.}
}
\description{
Tools for working with formula-like objects
}
