\name{skel}
\docType{data}
\alias{skel}
\title{Exploring relationships in body dimensions, skeletal measurements}
\description{
Skeletal diameter measurements, as well as height and gender, are
given for 507 physically active individuals - 247 men and 260 women.

This is a subset of the \code{\link{body}} data set.
}
\usage{skel}
\format{A data.frame containing 507 observations of 11 variables.}
\seealso{body}
\references{
Heinz, Peterson, Johnson and Kerk. "Exploring relationships in body 
dimensions". Journal of Statistics Education, 11(2). 
http://www.amstat.org/publications/jse/v11n2/datasets.heinz.html
}
\keyword{datasets}
