% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-step.R
\name{summary.stepArchetypes}
\alias{summary.stepArchetypes}
\title{Summary method for stepArchetypes object}
\usage{
\method{summary}{stepArchetypes}(object, ...)
}
\arguments{
\item{object}{A \code{stepArchetypes} object.}

\item{...}{Ignored.}
}
\value{
Undefined.
}
\description{
Summary method for stepArchetypes object
}
