% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_to_list}
\alias{table_to_list}
\title{Convert a Logical Table into a List}
\usage{
table_to_list(df, ...)
}
\arguments{
\item{df}{A \code{data.frame} object with logical values for each term}

\item{...}{Further arguments passed to or from other methods}
}
\description{
Takes a \code{data.frame} and uses the columns to generate a named list.
}
