% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{term}
\alias{term}
\alias{tx}
\alias{term.character}
\alias{term.formula}
\alias{term.default}
\alias{is_term}
\title{Vectorized terms using a \code{term_vctr}}
\usage{
tx(x = character(), ...)

term(x = character(), ...)

\method{term}{character}(
  x = character(),
  sides = character(),
  roles = character(),
  operations = character(),
  labels = character(),
  ...
)

\method{term}{formula}(x, roles = list(), labels = list(), ...)

\method{term}{default}(x, ...)

is_term(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{term()} or \code{tx()}: A character vector
\item For \code{is_term()}: An object to test
}}

\item{sides}{Left or right hand side of the equation}

\item{roles}{Specific role the variable plays within the formula}

\item{operations}{Modification of the term to be applied when combining with
data}

\item{labels}{Display-quality label describing the variable}
}
\description{
Vectorized terms using a \code{term_vctr}
}
