% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{terms}
\alias{terms}
\alias{term_rx}
\alias{term_rx.character}
\alias{term_rx.formula}
\alias{terms_rx.formula_rx}
\alias{term_rx.default}
\alias{trx}
\title{Term records}
\usage{
term_rx(x = character(), ...)

\method{term_rx}{character}(
  x = character(),
  side = character(),
  role = character(),
  group = character(),
  type = character(),
  operation = character(),
  label = character(),
  ...
)

\method{term_rx}{formula}(
  x = formula(),
  roles = list(),
  groups = list(),
  types = list(),
  labels = list(),
  ...
)

terms_rx.formula_rx(x, ...)

\method{term_rx}{default}(x, ...)

trx(x = character(), ...)
}
\arguments{
\item{x}{An object of the following types that can be coerced to a \code{term_rx} object
\itemize{
\item \code{character}
\item \code{formula}
}}

\item{side}{Left or right hand side of the equation}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure}: a predictor variable that serves as a primary or key
variable in the \eqn{Exposure ~ Outcome} relationship
}}

\item{group}{Grouping variable name for independent variables for modeling
terms together}

\item{operation}{Modification of the term to be applied when combining with
data}

\item{label}{Display-quality label describing the variable}
}
\value{
An object of class \code{formula_rx}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
