% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{trx}
\alias{trx}
\alias{term_rx}
\alias{term_rx.character}
\alias{term_rx.formula}
\alias{term_rx.data.frame}
\alias{term_rx.formula_rx}
\alias{term_rx.default}
\title{Term records}
\usage{
term_rx(x = unspecified(), ...)

\method{term_rx}{character}(
  x,
  side = character(),
  role = character(),
  group = character(),
  operation = character(),
  label = character(),
  description = character(),
  distribution = character(),
  subclass = character(),
  type = character(),
  ...
)

\method{term_rx}{formula}(
  x,
  roles = list(),
  groups = list(),
  labels = list(),
  descriptions = list(),
  distributions = list(),
  subclasses = list(),
  types = list(),
  ...
)

\method{term_rx}{data.frame}(x, ...)

\method{term_rx}{formula_rx}(x, ...)

\method{term_rx}{default}(x, ...)

trx(x = unspecified(), ...)
}
\arguments{
\item{x}{An object of the following types that can be coerced to a \code{term_rx}
object
\itemize{
\item \code{character}
\item \code{formula}
\item \code{data.frame}
}}

\item{side}{States the side of the formula the variable belongs on:
\itemize{
\item \strong{left}: For variables that are intended to be dependent
\item \strong{right}: For variables that are intended to be independent
\item \strong{meta}: For variables that are intended to explain relationships
between other variables, e.g. \emph{strata} or \emph{conditioning} variables
}}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure}: a predictor variable that serves as a primary or key
variable in the \eqn{Exposure ~ Outcome} relationship
\item \strong{outcome}: a outcome/dependent variable that serves as an individual
variable in the \eqn{Exposure ~ Outcome} relationship
\item \strong{covariate}: a predictor variable that is used to adjust/control for an
additional primary variable
\item \strong{mediator}: a predictor variable that is thought to be a causal
intermediary in the \eqn{Exposure -> Mediator -> Outcome} pathway
}}

\item{group}{Grouping variable name for independent variables for modeling
terms together}

\item{operation}{Modification of the term to be applied when combining with
data}

\item{label}{Display-quality label describing the variable}

\item{description}{Option for further descriptions or definitions needed for
the term, potentially part of a data dictionary}

\item{distribution}{If its associated with a data vector, describes the
distribution pattern of the original term}

\item{subclass}{Expected class of the variable, such as \code{character} or
\code{numeric}}

\item{type}{Type of the variable itself, e.g. ordinal, continuous,
dichotomous, etc}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
