% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe.R, R/terms.R
\name{axe.data.frame}
\alias{axe.data.frame}
\alias{axe.lm}
\alias{axe.glm}
\alias{axe.model_fit}
\alias{axe.formula_rx}
\alias{axe.default}
\alias{tx}
\alias{term_rex}
\title{Term records}
\usage{
axe.data.frame(x, ...)

axe.lm(
  x,
  role = list(),
  group = list(),
  label = list(),
  description = list(),
  distribution = list(),
  type = list(),
  subtype = list(),
  ...
)

axe.glm(
  x,
  role = list(),
  group = list(),
  label = list(),
  description = list(),
  distribution = list(),
  type = list(),
  subtype = list(),
  ...
)

axe.model_fit(
  x,
  role = list(),
  group = list(),
  label = list(),
  description = list(),
  distribution = list(),
  type = list(),
  subtype = list(),
  ...
)

axe.formula_rx(x, ...)

axe.default(x = unspecified(), ...)

term_rex(
  x,
  side = character(),
  role = character(),
  group = character(),
  operation = character(),
  label = character(),
  description = character(),
  distribution = character(),
  type = character(),
  subtype = character(),
  ...
)

tx(
  x,
  side = character(),
  role = character(),
  group = character(),
  operation = character(),
  label = character(),
  description = character(),
  distribution = character(),
  type = character(),
  subtype = character(),
  ...
)
}
\arguments{
\item{x}{An object of the following types that can be coerced to a \code{term_rex}
object. If it is an object that contains multiple terms, such as \code{formula},
the parameters are pluralized and should be contained via a list of
formulas. See details for further explanation.
\itemize{
\item \code{character}
\item \code{formula}
\item \code{lm}
\item \code{data.frame}
}}

\item{role}{Specific roles the variable plays within the formula. These are
of particular importance, as they serve as special terms that can effect
how a formula is interpreted. The options for roles are as below:
\itemize{
\item \strong{exposure}: a predictor variable that serves as a primary or key
variable in the \eqn{Exposure ~ Outcome} relationship
\item \strong{outcome}: a outcome/dependent variable that serves as an individual
variable in the \eqn{Exposure ~ Outcome} relationship
\item \strong{covariate}: a predictor variable that is used to adjust/control for an
additional primary variable
\item \strong{mediator}: a predictor variable that is thought to be a causal
intermediary in the \eqn{Exposure -> Mediator -> Outcome} pathway
}}

\item{group}{Grouping variable name for independent variables for modeling
terms together}

\item{label}{Display-quality label describing the variable}

\item{description}{Option for further descriptions or definitions needed for
the term, potentially part of a data dictionary}

\item{distribution}{If its associated with a data vector, describes the
distribution pattern of the original term}

\item{type}{The type of variable, either categorical (qualitative) or
continuous (quantitative)}

\item{subtype}{How the variable itself is more specifically subcategorized,
e.g. ordinal, continuous, dichotomous, etc}

\item{side}{States the side of the formula the variable belongs on:
\itemize{
\item \strong{left}: For variables that are intended to be dependent
\item \strong{right}: For variables that are intended to be independent
\item \strong{meta}: For variables that are intended to explain relationships
between other variables, e.g. \emph{strata} or \emph{conditioning} variables
}}

\item{operation}{Modification of the term to be applied when combining with
data}

\item{class}{Class of the variable itself, either expected or measured, such
as \code{character} or \code{numeric} or \code{factor}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Pluralized Arguments}{


For the arguments that would be dispatched for objects that are plural, e.g.
containing multiple terms such as a \code{formula} object, the input should be
wrapped within a \code{list()}.

For example, for the \strong{role} argument, it would be written:

\code{role = list(X ~ "exposure", Y ~ "outcome", M ~ "mediator", C ~ "covariate")}

This applies for all others plural objects and arguments.
}

