% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-weighted.R
\name{weightedArchetypes}
\alias{weightedArchetypes}
\title{Weighted archetypes}
\usage{
weightedArchetypes(data, k, weights = NULL, familyBlocks = list(), ...)
}
\arguments{
\item{data}{A numeric \eqn{n \times m} data matrix.}

\item{k}{The number of archetypes.}

\item{weights}{Data weights matrix.}

\item{familyBlocks}{Exchange predefined family blocks; see
\code{\link{archetypesFamily}}.}

\item{...}{Additional arguments for family blocks.}
}
\value{
An object of class \code{weightedArchetypes} and
  \code{\link{as.archetypes}}.
}
\description{
Weighted archetypes
}
\seealso{
Other archetypes: 
\code{\link{archetypesFamily}()},
\code{\link{archetypes_mod}()},
\code{\link{archetypes}},
\code{\link{as.archetypes}()},
\code{\link{robustArchetypes}}
}
\concept{archetypes}
