% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archetypes-xyplot.R
\name{xyplot}
\alias{xyplot}
\title{Two-dimensional plot.}
\usage{
xyplot(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Further arguments.}
}
\value{
Undefined.
}
\description{
Two-dimensional plot.
}
