\name{xyplot.robustArchetypes}
\alias{xyplot.robustArchetypes}
\title{Plot of two-dimensional data and robust archetypes.}
\usage{
\method{xyplot}{robustArchetypes}(x, y, ...)
}
\arguments{
  \item{x}{An \code{\link{archetypes}} object.}

  \item{y}{A matrix or data frame.}

  \item{...}{Arguments of
  \code{\link{xyplot.weightedArchetypes}} and
  \code{\link{xyplot.robustArchetypes}}}
}
\description{
Plot of two-dimensional data and robust archetypes.
}

