% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rss}
\alias{rss}
\alias{nparameters}
\alias{bestModel}
\alias{panorama}
\alias{pcplot}
\title{Defined generics}
\usage{
rss(object, ...)

nparameters(object, ...)

bestModel(object, ...)

panorama(object, ...)

pcplot(x, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Futher arguments}

\item{x}{An object.}
}
\description{
Generics defined by the archetypes package.
}
