\name{rss}
\alias{bestModel}
\alias{nparameters}
\alias{panorama}
\alias{pcplot}
\alias{rss}
\title{Defined generics}
\usage{
rss(object, ...)

nparameters(object, ...)

bestModel(object, ...)

panorama(object, ...)

pcplot(x, ...)
}
\arguments{
  \item{object}{An object}

  \item{...}{Futher arguments}

  \item{x}{An object.}
}
\description{
Generics defined by the archetypes package.

Return number of parameters

Return best model

Panorama

Parallel coordinates plot
}

