\name{archetypesFamily}
\alias{archetypesFamily}
\title{Archetypes family constructor.}
\usage{archetypesFamily(which=c("default", "ginv"))}
\description{Archetypes family constructor.}
\details{This function returns a problem solving brick for each of the different
conceptual parts of the algorithm. Currently, only the 'original' family
is supported.}
\value{A list containing a function for each of the different parts.}
\seealso{\code{\link{archetypes}}}
\arguments{\item{which}{The kind of archetypes family; currently ignored.}}
