\name{bestModel.stepArchetypes}
\alias{bestModel.repArchetypes}
\alias{bestModel.stepArchetypes}
\title{Return best model}
\usage{
  \method{bestModel}{stepArchetypes} (object, ...)

  \method{bestModel}{repArchetypes} (object, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{...}{Ignored}
}
\description{
  Return best model
}

