% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{checkers}
\alias{checkers}
\alias{is_term_rx}
\alias{is_formula_rx}
\title{Identification of objects in the \code{forks} package}
\usage{
is_term_rx(x)

is_formula_rx(x)
}
\arguments{
\item{x}{Confirmation of an object being of the following classes:
\itemize{
\item \code{term_rx}
\item \code{formula_rx}
}}
}
\description{
Identification of objects in the \code{forks} package
}
