\name{coef.archetypes}
\alias{coef.archetypes}
\title{Return coefficients}
\usage{
\method{coef}{archetypes}(object, type = c("alphas", "betas"), ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{type}{Return alpha or beta coefficients.}

  \item{...}{Ignored.}
}
\value{
Coefficient matrix.
}
\description{
Return coefficients
}

