% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{explode}
\alias{explode}
\title{Explode out a \code{list_of_formulas} object}
\usage{
explode(x, ...)
}
\value{
A \code{data.frame} object that is an explosion of the underlying
formulas, giving out the underlying terms, patterns, etc corresponding
formula
}
\description{
Explode out a \code{list_of_formulas} object
}
