% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{fit_fmls}
\alias{fit_fmls}
\title{Fitting a list of formulas}
\usage{
fit_fmls(object, .f, ..., data)
}
\arguments{
\item{object}{A \code{list_of_formulas} that can be fit by a modeling function,
such as \code{\link[stats:lm]{stats::lm()}}}
}
\value{
A list of model fits
}
\description{
Fitting a list of formulas
}
