\name{fitted.archetypes}
\alias{fitted.archetypes}
\title{Return fitted data}
\usage{
  \method{fitted}{archetypes} (object, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{...}{Ignored.}
}
\value{
  Matrix with approximated data.
}
\description{
  Returns the approximated data.
}

