% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{fmls}
\title{Formula Archetype}
\usage{
formula_archetype(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  strata = character(),
  pattern = character(),
  lineage = list(),
  order = "all",
  ...
)

fmls(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  strata = character(),
  pattern = character(),
  lineage = list(),
  order = "all",
  ...
)
}
\arguments{
\item{order}{Describes the requested order that formulas should be decomposed
into. The default is to return ALL formula decompositions. Options include any integer vector between \emph{1} and \emph{4}. If no value is given, all orders are returned.}
}
\description{
Formula Archetype
}
