% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{rx}
\alias{formula_frame}
\alias{formula_frame.term_vctr}
\alias{formula_frame.default}
\title{Vectorized formulas using a \code{formula_frame}}
\usage{
rx(x = term(), ...)

formula_frame(x = term(), ...)

\method{formula_frame}{term_vctr}(x = term(), pattern = character(), ...)

\method{formula_frame}{default}(x, ...)
}
\arguments{
\item{x}{\itemize{
\item For \code{formula_frame()} or \code{rx()}: A \code{term_vctr} object
}}

\item{pattern}{User-dependent methods to apply at the level of the formula at large}
}
\description{
Vectorized formulas using a \code{formula_frame}
}
