% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{formula_archetype.formula}
\alias{formula_archetype.term_archetype}
\alias{formula_archetype.default}
\alias{fm}
\title{Formula Archetype}
\usage{
formula_archetype(x = unspecified(), ...)

\method{formula_archetype}{formula}(x, tag = character(), ...)

\method{formula_archetype}{term_archetype}(x, ...)

\method{formula_archetype}{default}(x = unspecified(), ...)

fm(x = unspecified(), ...)
}
\description{
Formula Archetype
}
