% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{fmls}
\title{Formula Archetype}
\usage{
formula_archetype(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  strata = character(),
  pattern = character(),
  order = 2L:4L,
  ...
)

fmls(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  strata = character(),
  pattern = character(),
  order = 2L:4L,
  ...
)
}
\arguments{
\item{order}{Requested order that formulas are decomposed
into. The default is to return ALL formula decompositions. Options include
any integer inclusive from \emph{1} to \emph{4}. The default is \code{2L:4L}, which
includes every functional formula and its parent.}
}
\description{
Formula Archetype
}
