% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{formula_archetype.formula}
\alias{formula_archetype.default}
\alias{fx}
\title{Formula Archetype}
\usage{
formula_archetype(x = unspecified(), ...)

\method{formula_archetype}{formula}(x, ...)

\method{formula_archetype}{default}(x = unspecified(), ...)

fx(x = unspecified(), ...)
}
\description{
Formula Archetype
}
