% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{fmls}
\title{Formula Archetype}
\usage{
formula_archetype(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  family = character(),
  source = character(),
  pattern = character(),
  ...
)

fmls(
  x = unspecified(),
  role = list(),
  tier = list(),
  label = list(),
  family = character(),
  source = character(),
  pattern = character(),
  ...
)
}
\description{
Formula Archetype
}
