% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula}
\alias{formula}
\alias{formula_archetype}
\alias{formula_archetype.character}
\alias{formula_archetype.term_archetype}
\alias{formula_archetype.formula}
\alias{formula_archetype.default}
\alias{fmls}
\title{Formula Archetype}
\usage{
formula_archetype(x = unspecified(), ...)

\method{formula_archetype}{character}(
  x = character(),
  outcomes = character(),
  predictors = character(),
  exposures = character(),
  confounders = character(),
  mediators = character(),
  unknowns = character(),
  family = character(),
  source = character(),
  pattern = character(),
  level = character(),
  ...
)

\method{formula_archetype}{term_archetype}(x, ...)

\method{formula_archetype}{formula}(x, ...)

\method{formula_archetype}{default}(x = unspecified(), ...)

fmls(x = unspecified(), ...)
}
\description{
Formula Archetype
}
