% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{formula_vector}
\alias{formula_vector}
\alias{formula_vctr}
\alias{formula_vctr.term_rcrd}
\alias{formula_vctr.default}
\title{Formula vector}
\usage{
formula_vctr(x = term_rcrd(), ...)

\method{formula_vctr}{term_rcrd}(x = term_rcrd(), pattern = NA, ...)

\method{formula_vctr}{default}(x, ...)
}
\arguments{
\item{x}{Objects of the following types can be used as inputs
\itemize{
\item \code{term_rcrd}
}}

\item{...}{Arguments to be passed to or from other methods}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function defines a new modified \code{formula} class that has been
vectorized. It expands upon the functionality of formulas.
}
