% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{roles}
\alias{roles.term_archetype}
\alias{roles.script}
\alias{roles.list_of_formulas}
\alias{labels.term_archetype}
\alias{labels.script}
\alias{labels.list_of_formulas}
\alias{groups}
\alias{groups.term_archetype}
\alias{groups.script}
\alias{groups.list_of_formulas}
\title{Retrieval functions for \code{forks} classes}
\usage{
roles(x, ...)

\method{roles}{term_archetype}(x, ...)

\method{roles}{script}(x, ...)

\method{roles}{list_of_formulas}(x, ...)

\method{labels}{term_archetype}(x, ...)

\method{labels}{script}(x, ...)

\method{labels}{list_of_formulas}(x, ...)

groups(x, ...)

\method{groups}{term_archetype}(x, ...)

\method{groups}{script}(x, ...)

\method{groups}{list_of_formulas}(x, ...)
}
\description{
Retrieval functions for \code{forks} classes
}
