% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{get_terms}
\alias{get_terms.term_vctr}
\title{Get terms from prescribed formulas}
\usage{
get_terms(x, ...)

\method{get_terms}{term_vctr}(x, side = "both", ...)
}
\description{
Get terms from prescribed formulas
}
