% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{lhs}
\alias{rhs}
\alias{rhs.term_rcrd}
\alias{lhs.term_rcrd}
\alias{rhs.formula}
\alias{lhs.formula}
\alias{getComponent}
\alias{getComponent.term_rcrd}
\title{Get terms from prescribed formulas}
\usage{
lhs(x, ...)

rhs(x, ...)

\method{rhs}{term_rcrd}(x, ...)

\method{lhs}{term_rcrd}(x, ...)

\method{rhs}{formula}(x, tidy = FALSE, ...)

\method{lhs}{formula}(x, ...)

getComponent(x, ...)

\method{getComponent}{term_rcrd}(x, part, filter_id = NULL, filter_val = NULL, ...)
}
\description{
Get terms from prescribed formulas
}
