% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{roles}
\alias{roles.term_archetype}
\alias{roles.script}
\alias{get_terms}
\alias{labels.term_archetype}
\alias{labels.script}
\alias{labels.list_of_formulas}
\alias{tiers}
\alias{tiers.term_archetype}
\alias{tiers.script}
\alias{tiers.list_of_formulas}
\title{Retrieval functions for \code{forks} classes}
\usage{
roles(x, ...)

\method{roles}{term_archetype}(x, ...)

\method{roles}{script}(x, ...)

get_terms(x, field = NA, value = NA)

\method{labels}{term_archetype}(object, ...)

\method{labels}{script}(object, ...)

\method{labels}{list_of_formulas}(object, ...)

tiers(x, ...)

\method{tiers}{term_archetype}(x, ...)

\method{tiers}{script}(x, ...)

\method{tiers}{list_of_formulas}(x, ...)
}
\description{
Retrieval functions for \code{forks} classes
}
