% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{roles}
\alias{roles.term}
\alias{roles.rx}
\alias{roles.formula_list}
\alias{labels.term}
\alias{labels.rx}
\alias{labels.formula_list}
\alias{groups}
\alias{groups.term}
\alias{groups.rx}
\alias{groups.formula_list}
\title{Retrieval functions for \code{forks} classes}
\usage{
roles(x, ...)

\method{roles}{term}(x, ...)

\method{roles}{rx}(x, ...)

\method{roles}{formula_list}(x, ...)

\method{labels}{term}(x, ...)

\method{labels}{rx}(x, ...)

\method{labels}{formula_list}(x, ...)

groups(x, ...)

\method{groups}{term}(x, ...)

\method{groups}{rx}(x, ...)

\method{groups}{formula_list}(x, ...)
}
\description{
Retrieval functions for \code{forks} classes
}
