% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{roles}
\alias{roles.term_rx}
\alias{roles.formula_rx}
\alias{roles.list_of_formulas}
\alias{labels.term_rx}
\alias{labels.formula_rx}
\alias{labels.list_of_formulas}
\alias{labels.fmls}
\alias{groups}
\alias{groups.term_rx}
\alias{groups.formula_rx}
\alias{groups.list_of_formulas}
\title{Retrieval functions for \code{forks} classes}
\usage{
roles(x, ...)

\method{roles}{term_rx}(x, ...)

\method{roles}{formula_rx}(x, ...)

\method{roles}{list_of_formulas}(x, ...)

\method{labels}{term_rx}(x, ...)

\method{labels}{formula_rx}(x, ...)

\method{labels}{list_of_formulas}(x, ...)

\method{labels}{fmls}(x, ...)

groups(x, ...)

\method{groups}{term_rx}(x, ...)

\method{groups}{formula_rx}(x, ...)

\method{groups}{list_of_formulas}(x, ...)
}
\description{
Retrieval functions for \code{forks} classes
}
