% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getters}
\alias{getters}
\alias{get_terms}
\alias{get_roles}
\title{Get terms from prescribed formulas}
\usage{
get_terms(x, side = "both", ...)

get_roles(x, role = "all", ...)
}
\description{
Get terms from prescribed formulas
}
