% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_of_formulas}
\alias{list_of_formulas}
\alias{list_of_formulas.formula_rcrd}
\alias{list_of_formulas.default}
\title{Prescribed Formula Lists}
\usage{
list_of_formulas(x, ...)

\method{list_of_formulas}{formula_rcrd}(x = formula_vctr(), pattern = character(), ...)

\method{list_of_formulas}{default}(x, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super class that combines both the \code{list} class (and its derivative \code{list_of}) and the \code{formula} class.
}
