% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_lists.R
\name{list_of_formulas}
\alias{list_of_formulas}
\alias{list_of_formulas.formula_script}
\alias{list_of_formulas.list}
\alias{list_of_formulas.default}
\alias{fmls}
\title{Prescribed Formula Lists}
\usage{
list_of_formulas(x = unspecified(), ...)

\method{list_of_formulas}{formula_script}(
  x,
  tag = deparse1(substitute(x)),
  pattern = character(),
  strata = character(),
  ...
)

\method{list_of_formulas}{list}(
  x,
  tag = deparse1(substitute(x)),
  pattern = character(),
  strata = character(),
  ...
)

\method{list_of_formulas}{default}(x = list(), ...)

fmls(x = unspecified(), ...)
}
\arguments{
\item{pattern}{This is the expansion pattern used to decide how the
covariates will incorporated into the formulas. The options are
\code{c("direct", "sequential", "parallel")}. See the details for further
explanation.
\itemize{
\item \strong{direct}: the covariates will all be included in each formula
\item \strong{sequential}: the covariates will be added sequentially, one by one, or
by groups, as indicated
\item \strong{parallel}: the covariates or groups of covariates will be placed in
parallel
}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function introduces a super  that combines both the \code{list} class (and
its derivative \code{list_of}) and the \code{formula} class.
}
\section{Patterns}{


The expansion pattern allows for instructions on how the covariates should be
included in different formulas. Below, assuming that \emph{x1}, \emph{x2}, and \emph{x3} are
covariates...

\deqn{y ~ x1 + x2 + x3}

\strong{Direct}:

\deqn{y ~ x1 + x2 + x3}

\strong{Seqential}:

\deqn{y ~ x1}
\deqn{y ~ x1 + x2}
\deqn{y ~ x1 + x2 + x3}

\strong{Parallel}:

\deqn{y ~ x1}
\deqn{y ~ x2}
\deqn{y ~ x3}
}

