% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_table}
\alias{list_to_table}
\title{Convert a List into a Table}
\usage{
list_to_table(x, id = "terms", val = "ops", ...)
}
\arguments{
\item{x}{A named \code{list} object}

\item{id}{Name of column that contains terms}

\item{val}{Name of column that contains specific values}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{data.frame} object
}
\description{
Expands a list of character vectors into a tidy, logic table.
}
