% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{model_archetype}
\alias{model_archetype.lm}
\alias{model_archetype.glm}
\alias{model_archetype.model_fit}
\alias{model_archetype.list}
\alias{model_archetype.default}
\alias{md}
\title{Model archetype}
\usage{
model_archetype(x = unspecified(), ...)

\method{model_archetype}{lm}(
  x = unspecified(),
  name = deparse(substitute(x)),
  label = character(),
  description = character(),
  term_labels = list(),
  term_roles = list(),
  ...
)

\method{model_archetype}{glm}(
  x = unspecified(),
  name = deparse(substitute(x)),
  label = character(),
  description = character(),
  term_labels = list(),
  term_roles = list(),
  ...
)

\method{model_archetype}{model_fit}(
  x = unspecified(),
  name = deparse(substitute(x)),
  label = character(),
  description = character(),
  term_labels = list(),
  term_roles = list(),
  ...
)

\method{model_archetype}{list}(
  x = unspecified(),
  label = character(),
  description = character(),
  term_labels = list(),
  term_roles = list(),
  ...
)

\method{model_archetype}{default}(x = unspecified(), ...)

md(x = unspecified(), ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
