\name{nparameters.archetypes}
\alias{nparameters.archetypes}
\alias{nparameters.repArchetypes}
\alias{nparameters.stepArchetypes}
\title{Return number of archetypes}
\usage{
\method{nparameters}{archetypes}(object, ...)

\method{nparameters}{stepArchetypes}(object, ...)

\method{nparameters}{repArchetypes}(object, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{...}{Ignored.}
}
\value{
Number of archetypes.
}
\description{
Return number of archetypes
}

