\docType{methods}
\name{parameters,archetypes-method}
\alias{parameters,archetypes-method}
\alias{parameters,repArchetypes-method}
\alias{parameters,stepArchetypes-method}
\alias{parameters-methods}
\title{Return fitted archetypes}
\usage{
\S4method{parameters}{archetypes}(object, ...)

\S4method{parameters}{stepArchetypes}(object, ...)

\S4method{parameters}{repArchetypes}(object, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{...}{Ignored.}
}
\value{
Matrix with \eqn{k} archetypes.
}
\description{
Return fitted archetypes
}

