\name{parameters}
\alias{parameters}
\alias{parameters-methods}
\alias{parameters,archetypes-method}
\alias{parameters,repArchetypes-method}
\alias{parameters,stepArchetypes-method}
\title{Return fitted archetypes}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{...}{Ignored.}
}
\value{
  Matrix with \eqn{k} archetypes.
}
\description{
  Return fitted archetypes
}

