% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{rebuild}
\alias{rebuild}
\title{Rebuild a formula-based object as a data frame}
\usage{
rebuild(x, ...)
}
\value{
A \code{data.frame} object that is an explosion of the underlying object,
giving out the underlying terms, patterns, etc, corresponding to the
initial properties of the object formula
}
\description{
Rebuild a formula-based object as a data frame
}
