\name{rss.archetypes}
\alias{rss.archetypes}
\alias{rss.repArchetypes}
\alias{rss.stepArchetypes}
\title{Return residual sum of squares}
\usage{
\method{rss}{archetypes}(object, type = c("scaled", "single", "global"), ...)

\method{rss}{stepArchetypes}(object, ...)

\method{rss}{repArchetypes}(object, ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{type}{Return scaled, single or global RSS.}

  \item{...}{Ignored.}
}
\value{
Residual sum of squares.
}
\description{
Return residual sum of squares
}

