% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setters}
\alias{setters}
\alias{setGroups}
\alias{setGroups.term_rx}
\title{Set components of terms and formulas}
\usage{
setGroups(x, ...)

\method{setGroups}{term_rx}(x, groups, ...)
}
\value{
A modified
}
\description{
Set components of terms and formulas
}
