% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setters}
\alias{setters}
\alias{setRoles}
\alias{setGroups}
\alias{setLabels}
\title{Set components of terms and formulas}
\usage{
setRoles(x, roles, ...)

setGroups(x, groups, ...)

setLabels(x, labels, ...)
}
\value{
A modified
}
\description{
Set components of terms and formulas
}
