\name{summary.stepArchetypes}
\alias{summary.stepArchetypes}
\title{Summary method for stepArchetypes object}
\usage{
\method{summary}{stepArchetypes}(object, ...)
}
\arguments{
  \item{object}{A \code{stepArchetypes} object.}

  \item{...}{Ignored.}
}
\value{
Undefined.
}
\description{
Summary method for stepArchetypes object
}

