% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{term}
\alias{term}
\alias{tx}
\alias{term.character}
\alias{term.formula}
\alias{term.default}
\alias{is_term}
\title{\code{term} vector}
\usage{
tx(x, ...)

term(x, ...)

\method{term}{character}(
  x = character(),
  sides = character(),
  roles = character(),
  operations = character(),
  labels = character()
)

\method{term}{formula}(x, roles = list(), labels = list())

\method{term}{default}(x, ...)

is_term(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{term()} or \code{tx()}: A character vector
\item For \code{is_term()}: An object to test
}}
}
\description{
\code{term} vector
}
