% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{term_record}
\alias{term_record}
\alias{term_rcrd}
\alias{term_rcrd.character}
\alias{term_rcrd.formula}
\alias{term_rcrd.default}
\title{Term records}
\usage{
term_rcrd(x = character(), ...)

\method{term_rcrd}{character}(
  x = character(),
  side = character(),
  role = character(),
  group = character(),
  type = character(),
  operation = character(),
  label = character(),
  ...
)

\method{term_rcrd}{formula}(
  x = formula(),
  roles = list(),
  groups = list(),
  types = list(),
  labels = list(),
  ...
)

\method{term_rcrd}{default}(x, ...)
}
\arguments{
\item{x}{An object of the following types that can be coerced to a \code{term_rcrd} object
\itemize{
\item \code{character}
\item \code{formula}
}}

\item{roles}{Specific role the variable plays within the formula}

\item{labels}{Display-quality label describing the variable}

\item{sides}{Left or right hand side of the equation}

\item{operations}{Modification of the term to be applied when combining with
data}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
