% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updates}
\alias{updates}
\alias{update.term_archetype}
\alias{update.formula_script}
\alias{add}
\alias{add.formula_script}
\alias{add.term_archetype}
\title{Update Prescriptions}
\usage{
\method{update}{term_archetype}(object, parameters, ...)

\method{update}{formula_script}(object, parameters, ...)

add(object, ...)

\method{add}{formula_script}(object, parameters, ...)

\method{add}{term_archetype}(object, parameters, ...)
}
\value{
An object of the original class
}
\description{
These are a variety of functions to help update and modify objects from the
\code{{forks}} package.
}
