% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updates}
\alias{updates}
\alias{update.formula_rx}
\alias{add}
\title{Updating prescribed formulas}
\usage{
\method{update}{formula_rx}(object, parameters, ...)

add(object, parameters, ...)
}
\value{
An object of class \code{formula_rx}
}
\description{
Updating prescribed formulas
}
