% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updates}
\alias{updates}
\alias{update.term}
\alias{update.rx}
\alias{add}
\alias{add.rx}
\alias{add.term}
\title{Update Prescriptions}
\usage{
\method{update}{term}(object, parameters, ...)

\method{update}{rx}(object, parameters, ...)

add(object, ...)

\method{add}{rx}(object, parameters, ...)

\method{add}{term}(object, parameters, ...)
}
\value{
An object of the original class
}
\description{
These are a variety of functions to help update and modify objects from the
\code{{forks}} package.
}
