\name{weights.archetypes}
\alias{weights.archetypes}
\title{Return weights}
\usage{
\method{weights}{archetypes}(object, type = c("weights", "reweights"), ...)
}
\arguments{
  \item{object}{An \code{archetypes} object.}

  \item{type}{Return global weights (weighted archetypes)
  or weights calculated during the iterations (robust
  archetypes).}

  \item{...}{Ignored.}
}
\value{
Vector of weights.
}
\description{
Return weights
}

